feat!(website): make metadata docs generally available
