+++ website/src/components/Submission/metadataFormatDocsUrl.ts
@@ -1,0 +1,2 @@
+// This variable is in a separate file as it is overwritten by Pathoplexus.
+export const metadataFormatDocsUrl = '../metadataformat';
