--- website/src/pages/docs/concepts/metadataformat.mdx
@@ -1,3 +1,0 @@
-## Metadata Format
-
-Fill this in with tips for your loculus instance.
